gen: Fix ICH SPI driver to work for write transactions.
There have been a couple of problems with the driver:

- the commands which do not involve data transfer (WREN most
notably) were not being issued.

- the write page command was sent trying to send address as data,
which was causing the 3 bytes of the command to overflow to the
next cycle

This change fixes these problems and also refactors the driver
such that the status polling with timeout is implemented once and
used from different places.

BUG=chromium-os:20105
TEST=manual

. this change was tested combined with other modifications; The
ability to read and save the u-boot environment in an SPI
flash was demonstrated.

Change-Id: I12da0363eaf9ae193d1d29a84a9a2b48cbd58fd8
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/7526
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Gabe Black <gabeblack@chromium.org>