gen: Add an gnvs-vboot-table stub entry to chromeos-config on x86
crossystem related information is passed in to chromeos using ACPI on x86. On
coreboot systems, the location of the table this information actually comes
from will be injected into the device tree into a stub entry called
gnvs-vboot-table which is in the chromeos-config node. u-boot will fill out
this information, and when the kernel calls the appropriate ACPI methods the
right things will happen.

BUG=chrome-os-partner:5944
BUG=chrome-os-partner:5961
BUG=chrome-os-partner:5962
TEST=Booted successfully this change.

Change-Id: Ieca2a7627935dbe531ef9a4dc9c5b0bc0f97ad7f
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/7851
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Commit-Ready: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>