gen: Remove manual configuration of the serial console from the device tree
Coreboot is going to set up the serial configuration in the device tree at run
time, and this manual configuration would get in the way.

BUG=chrome-os-partner:4901
TEST=Booted on Stumpy and Alex with the change in coreboot that makes it
configure the serial device. Booted on Alex with the external serial card
removed and its driver disabled in coreboot and no other change in u-boot.

Change-Id: Ibdf2590722c822ce663a62516a5d1c374e8b1293
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/8473
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>