lost: split: cros: x86: Process all cbmem table references consistently.
Rebased-from: b446c83
Lost-file: arch/x86/cpu/coreboot/tables.c

A few CBMEM table entries are included in coreboot to let u-boot
discover the locations of those entries. They all are passed
using the same structure within coreboot table. This patch makes
use of that structure for all communicated entries (timestam,
console and MRC cache).

It does not make sense to keep types of pointers in the sysinfo
table, as nobody but the respective users of those fields use the
types. Let's keep the pointers as void *, this also allows to
reduce the amount of include files in systinfo.h and hide the
timestamp structures in timestamp.c.

BUG=chrome-os-partner:4200
TEST=manual

. build the new firmware and program it on a stumpy
. restart the machine
. after it comes up to ChromeOS, run the cbmem.py utility

Observe that timstamp values are displayed, and there is an entry
with index 1000 (added by u-boot to the coreboot timestamp
table).

Change-Id: Icb808145c4c62cee939eceb3d5bf5afb3bcd00d9
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/8711
Reviewed-by: Stefan Reinauer <reinauer@google.com>