Allow arch directory to contain .lds without requiring Makefile
The Makefile for a CPU is in arch/($ARCH)/cpu/$(CPU). We want to support
having an .lds file in arch/$(ARCH)/cpu without requiring an additional
Makefile there. This change makes it clear that we expect a Makefile in
the same directory as the link script except in this case.

Series-changes: 2
- Add a comment explaining all the ldscript checks

Change-Id: I203ad34ff63177320f0d8da5428dc53431c9a21b