timer: add microsecond boot func
Define timer_get_boot_us() which returns the number of microseconds
since boot. If undefined then we use get_timer() * 1000.

We can fit this in a 32-bit register which keeps everyone happy on
the efficiency side. It will wrap around after about an hour. If we
are still looking at it after an hour then we had better not be
timing the boot.

There is some scope to join bootstage and post, but for now they are
separate things with separate CONFIGs, and you don't have to enable
one to get the other.

Series-changes: 3
- Rename timer patch to remove bootstage: tag

Change-Id: Ifc0fab4921cb0a6188d1ea1adf705fb343377e26
(cherry picked from commit 6728b1e91b861ff9219a1421a59c2dd02d4d79cb)