upstream: tegra: usb: fdt: Add additional device tree definitions for USB ports
This adds clock references to the USB part of the device tree for U-Boot,
and marks USB1 as supporting legacy mode (which we disable in the driver).

The USB timing information may vary between boards sometimes, but for
now we hard-code it in C. This is because all current T2x boards use
the same values, we will deal with T3x later and we first need to agree
on the format for this timing information in the fdt and may in fact
decide that it has no place there.

Series-changes: 5
- Change device tree comment style from // to /* */
- Add nvidia,has-legacy-mode property per review comments
- Add dr_mode property to control host/device/otg mode

Series-changes: 6
- Remove dr_mode properties from SOC .dtsi file and move to boards

Change-Id: I3c0ca6f0de43038449a829be6976612767b23b64
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/18740
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>