upstream: tegra: i2c: Enable I2C on Seaboard
This enables I2C on Seaboard.

Series-to: u-boot
Series-cc: tom, stephen, linux-tegra
Series-version: 5
Cover-letter:
tegra: Add I2C driver and associated parts
This series brings in an I2C driver for Tegra which can be
configured by a flat device tree.

It supports 8- and 16-bit addresses and both the normal I2C ports and
the DVC port (for controlling the power management unit (PMU)).

Recent Linux bindings are used, based on example .dts files found in
branch for-3.3/dt at:

git://git.kernel.org/pub/scm/linux/kernel/git/olof/tegra.git

(I could not find the actual binding documentation to include here)

Note: There is still an open question on I2C2 pinmux. The idea of an
'nvidia,pinmux-config' settings did not find favour, so we have no way
of passing the correct config to funcmux for this port. Therefore we have
disabled it for now. The impact on Seaboard is that the TPM and battery
charger cannot be accessed.

I am sending this series to the linux-tegra list also, and will do so for
future series also. This is because anything to do with fdt also affects
the kernel bindings. Please advise if this is not useful.
END
Series-changes: 3
- Use CONFIG_TEGRA_I2C instead of CONFIG_TEGRA2_I2C

Change-Id: I65f89c230d3ff1025ff208ab07c546c5b4c698d8
Reviewed-on: https://gerrit.chromium.org/gerrit/18719
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>