fdt: Add fdtdec_find_aliases() to deal with alias nodes
Stephen Warren pointed out that we should use nodes whether or not they
have an alias in the /aliases section. The aliases section specifies the
order so far as it can, but is not essential. Operating without alisses
is useful when the enumerated order of nodes does not matter (admittedly
rare in U-Boot).

This is considerably more complex, and it is important to keep this
complexity out of driver code. This patch creates a function
fdtdec_find_aliases() which returns an ordered list of node offsets
for a particular compatible ID, taking account of alias nodes.

Series-changes: 2
- Rename function to fdtdec_find_aliases_for_id()
- Improve alias checking algorithm to reduce run time
- Skip nodes marked as disabled
- Allow gaps in the list returned to the caller

Change-Id: Id00d0a827e34206e46e972322bafd34251a582e7