Introduce generic link symbol file
We create a separate header file for link symbols defined by the link
scripts. It is helpful to have these all in one place and try to
make them common across architectures. Since Linux already has a similar
file, we bring this in even though many of the symbols there are not
relevant to us.

Change-Id: Idd515433426307815d44d2b7cd87fe6f5f851a17